requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-multipart==0.0.9
psycopg2-binary==2.9.9
SQLAlchemy==2.0.35
pgvector==0.3.2
sentence-transformers==2.7.0
torch==2.3.1
numpy==1.26.4
pymupdf==1.24.9
python-docx==1.1.2
langdetect==1.0.9
groq==0.10.0
# Nếu dùng OpenAI:
# openai>=1.40.0
